Get realtime demo working with TFRecords.
I've tested it with my trained model and it works pretty well.

+++ gaze_vis.py
@@ -7,17 +7,25 @@ class GazeVis(object):
   def __init__(self, window_width=1920, window_height=1070):
     # Create canvas window.
     self.__canvas = oc.Canvas(window_width, window_height)
-
-    # Draw the initial point.
-    self.__dot = oc.Circle(self.__canvas, (0, 0), 15, fill="red")
+    self.__dot = None
 
   def visualize_point(self, point):
     """ Visualizes the specified point.
     Args:
       point: The x and y coordinates of the point to visualize. """
     new_x, new_y = point
+    if self.__dot is None:
+      # Make a new one if we deleted it.
+      self.__dot = oc.Circle(self.__canvas, (new_x, new_y), 15, fill="red")
+      return
+
     self.__dot.set_pos(new_x, new_y)
 
   def hide_dot(self):
     """ Hides the dot, in the case of no prediction. """
+    if self.__dot is None:
+      # Already hidden.
+      return
+
     self.__dot.delete()
+    self.__dot = None
